requirements.txt
numpy
pandas
streamlit
scikit-learn
pydantic
matplotlib
scipy
python-binance
ipykernel